move writeArray() into buffer v2 from base
